refactor: TeamInvestmentCompletedEventDTO 가격관련 필드 Int -> Long으로 수정

+++ src/main/kotlin/team/mozu/dsm/adapter/in/team/dto/TeamInvestmentCompletedEventDTO.kt
@@ -6,7 +6,7 @@ data class TeamInvestmentCompletedEventDTO(
     val teamId: UUID,
     val teamName: String?,
     val curInvRound: Int,
-    val totalMoney: Int,
-    val valuationMoney: Int,
+    val totalMoney: Long,
+    val valuationMoney: Long,
     val profitNum: Double
 )
